app/requirements.txt
python-dotenv
flask
flask-sqlalchemy
flask-bcrypt
flask-cors
flask-session
redis
PyJWT
